arm64/dts: caimito: Define camera PMIC supplies.
This serves as the foundation of the second attempt for enabling camera
buck-boost converter FPWM mode.

This CL defines regulators that are responsible of some camera supply
LDOs and PMIC: PP1350_CAM and PP3200_CAM_BB. Instead of having SLG51002
control the enablement of the regulators through GPIO, we gave the
regulator those GPIO pins instead. PMIC VOUT supplies will depend on
the new regulators instead, making sure that corresponding supplies will
be turned on before accessing the VOUT supply regulators.

By having PMIC VOUT supplies to depend on the supply regulators, we can
control the PP3200_CAM_BB regulator to be FPWM mode from the driver
through REGULATOR_FAST_MODE (to be done on next change).

We also include these new regulators into each of the camera power
sequences. This is to follow the FCAM VDIG LDO sequencing requirement
(which the LDO itself is supplied from PP1350_CAM and PP3200_CAM_BB).
We also need to put the sequence in all cameras because we want to make
sure that the sequence is always followed, regardless of which camera we
turn on (for example, dual camera stream where we turn on rear camera
and then front camera, rear camera may only turn on PP1350_CAM and by
the time the front camera is turned on, the sequencing requirement is
violated.)

Overall, there should be no side effect on the hardware behavior.

Bug: 293455894
Bug: 296883244
Test: Boot, Camera preview
Change-Id: Ia564566aaa6412db6900c79fc6ab99289dca0d8c
Signed-off-by: Firman Prayoga <firmanhp@google.com>